requirements.txt
RPi.GPIO==0.6.3
GPIOSimulator==0.1